Added some more options for setup
File `configure.pl` uses the Template Toolkit
from Perl to configure file `config.tt`. Should
probably change the ".tt" suffix on `config.tt`
to be the same as the actual target file type.
This is because ".tt" implies it is a file used
with `vimplate`. Here we won't be directly using
`vimplate`.